requirements.txt
discord.py
requests
PIL
discord.ext
matplotlib
O365
Pillow
tzlocal==2.1
youtube_dl
asyncio
time
io
PyNaCl